Feature CRUD upload image

+++ backend/src/controllers/question.controller.ts
@@ -1,15 +1,35 @@
 import { Request, Response } from "express";
+import { questionImage } from "../models/image.model";
 import { QuestionModel } from "../models/question.model";
+import { uploadImage } from "../types/image";
 import { CreateQuestion, UpdateQuestion } from "../types/question";
 
 const questionModel = new QuestionModel();
+const imageService = new questionImage();
 
 class QuestionController {
-  async createQuestion(req: Request<null, null, CreateQuestion>, res: Response) {
-    const question = req.body;
+  async createQuestion(
+    req: Request<{}, {}, CreateQuestion>, // Correct typing for POST (no URL params)
+    res: Response,
+  ): Promise<void> {
     try {
-      const newQuiz = await questionModel.createQuestion(question);
-      res.status(201).json(newQuiz);
+      const question = req.body;
+      if (req.file) {
+        if (req.file) {
+          const imageInfo: uploadImage = {
+            fileBuffer: req.file?.buffer,
+            mimeType: req.file?.mimetype,
+          };
+          const uploadImage = await imageService.uploadImage(imageInfo);
+          question.picture = uploadImage?.imageUrl;
+        }
+      }
+      try {
+        const newQuiz = await questionModel.createQuestion(question);
+        res.status(201).json(newQuiz);
+      } catch (error) {
+        console.error(`Error ${error}`);
+      }
     } catch (error) {
       console.error("Error creating question", error);
       res.status(500).json({ error: "Error creating question" });
@@ -19,7 +39,25 @@ class QuestionController {
   async updateQuestion(req: Request<{ id: string }, null, UpdateQuestion>, res: Response) {
     const id = req.params.id;
     const question = req.body;
+    let newImageUrl;
     try {
+      if (req.file) {
+        const imageInfo: uploadImage = {
+          fileBuffer: req.file?.buffer,
+          mimeType: req.file?.mimetype,
+        };
+        const uploadImage = await imageService.uploadImage(imageInfo);
+        newImageUrl = uploadImage?.imageUrl;
+        const existingQuestion = await questionModel.getQuestionById(id);
+        // delete the old image
+        if (existingQuestion) {
+          if (existingQuestion.picture !== question.picture) {
+            await imageService.deleteImage(existingQuestion.picture);
+          }
+        }
+        question.picture = newImageUrl;
+      }
+
       const updatedQuiz = await questionModel.updateQuestion(id, question);
       res.status(200).json(updatedQuiz);
     } catch (error) {
@@ -58,6 +96,16 @@ class QuestionController {
   async deleteQuestion(req: Request, res: Response) {
     const id = req.params.id;
     try {
+      const question = await questionModel.getQuestionById(id);
+      let imageUrl = "";
+      if (question) {
+        imageUrl = question.picture;
+      }
+      const deleted = await imageService.deleteImage(imageUrl);
+      if (!deleted) {
+        res.status(404).json({ message: "Could not delete the image" });
+        return;
+      }
       const deletedQuestion = await questionModel.deleteQuestion(id);
       res.status(200).json(deletedQuestion);
     } catch (error) {
